Break out stock info card generation to utils.
